Moved the rating field component to its own file.
