+++ meteor-app/imports/ui/components/form-stuff/field/field.tsx
@@ -1,34 +1,8 @@
 import React from "react";
 import * as Formik from "formik";
 import { TextField } from "formik-material-ui";
-import { Rating } from "@material-ui/lab";
-import styled from "styled-components";
 
-const RatingLayout = styled.div`
-	display: flex;
-	flex-direction: row-reverse;
-	justify-content: left;
-
-	margin-top: 16px;
-`;
-
-const Legend = styled.div`
-	margin-left: 10px;
-`;
-
-type RatingFieldProps = Formik.FieldProps<any> & {
-	label: string;
-};
-
-const RatingField = ({
-	label,
-	field: { name, value, onChange },
-}: RatingFieldProps) => (
-	<RatingLayout>
-		<Legend>{label}</Legend>
-		<Rating name={name} value={value} onChange={onChange} />
-	</RatingLayout>
-);
+import RatingField from "./rating-field";
 
 const FieldInner: React.ComponentType<any> = ({ type, ...restProps }) => {
 	if (type === "rating") {
